团签账单

+++ app/model/group_user/group_user_entity.go
@@ -12,7 +12,7 @@ import (
 
 // Entity is the golang structure for table group_user.
 type Entity struct {
-    Id        uint        `orm:"id,primary" json:"id"`         //           
+    Id        uint64      `orm:"id,primary" json:"id"`         //           
     GroupId   uint        `orm:"groupId"    json:"group_id"`   // 团队ID    
     UserId    uint64      `orm:"userId"     json:"user_id"`    // 用户ID    
     DeletedAt *gtime.Time `orm:"deletedAt"  json:"deleted_at"` // 注销时间
