+++ app/model/internal/user_biz.go
@@ -16,8 +16,8 @@ type UserBiz struct {
 	CityId       uint        `orm:"cityId"       json:"cityId"`       // 城市ID
 	ShopId       uint        `orm:"shopId"       json:"shopId"`       // 门店ID
 	GoroupId     uint        `orm:"goroupId"     json:"goroupId"`     // 团签ID
-	GoroupUserId uint        `orm:"goroupUserId" json:"goroupUserId"` // 团签用户ID
-	Type         uint        `orm:"type"         json:"type"`         // 业务类型: 1新签 2换电 3寄存 4退租
+	GoroupUserId uint64      `orm:"goroupUserId" json:"goroupUserId"` // 团签用户ID
+	Type         uint        `orm:"type"         json:"type"`         // 业务类型: 1新签 2换电 3寄存 4恢复计费 5退租
 	BatteryType  string      `orm:"batteryType"  json:"batteryType"`  // 电池型号
 	CreatedAt    *gtime.Time `orm:"createdAt"    json:"createdAt"`    // 扫码时间
 	UpdatedAt    *gtime.Time `orm:"updatedAt"    json:"updatedAt"`    //
