+++ app/model/user_biz/user_biz_entity.go
@@ -18,7 +18,7 @@ type Entity struct {
     CityId       uint        `orm:"cityId"       json:"city_id"`        // 城市ID                                       
     ShopId       uint        `orm:"shopId"       json:"shop_id"`        // 门店ID                                       
     GoroupId     uint        `orm:"goroupId"     json:"goroup_id"`      // 团签ID                                       
-    GoroupUserId uint        `orm:"goroupUserId" json:"goroup_user_id"` // 团签用户ID                                   
+    GoroupUserId uint64      `orm:"goroupUserId" json:"goroup_user_id"` // 团签用户ID                                   
     Type         uint        `orm:"type"         json:"type"`           // 业务类型: 1新签 2换电 3寄存 4恢复计费 5退租  
     BatteryType  string      `orm:"batteryType"  json:"battery_type"`   // 电池型号                                     
     CreatedAt    *gtime.Time `orm:"createdAt"    json:"created_at"`     // 扫码时间
